Move API mocking to a dedicated file
